add local storage after unload the page

+++ src/ts/DataStorage.ts
@@ -2,26 +2,31 @@ import { TPokemonsData } from "./types";
 import { POKEMONS_LIST_KEY } from "./utlites/constantVariables";
 
 export class DataStorage {
-  static initEvent(pokemonData: TPokemonsData) {
-    DataStorage.loadDataEvent(pokemonData);
+  static async initEvent(pokemonData: TPokemonsData) {
+    await DataStorage.loadDataEvent(pokemonData);
     DataStorage.saveDataEvent(pokemonData);
   }
 
-  static loadDataEvent(pokemonData: TPokemonsData) {
+  static async loadDataEvent(pokemonData: TPokemonsData) {
     const localStorageData = DataStorage.checkLocalStorageExist();
 
-    if (!localStorageData) return;
-    pokemonData.setItems(JSON.parse(localStorageData));
-    DataStorage.removeLocalStorage();
+    if (localStorageData) {
+      pokemonData.setItems(JSON.parse(localStorageData));
+      DataStorage.removeLocalStorage();
+    } else await pokemonData.fetchPokemonsListDetails(1, 905);
   }
 
   static saveDataEvent(pokemonData: TPokemonsData) {
     window.addEventListener("unload", () => {
-      if (!DataStorage.checkLocalStorageExist())
+      if (
+        !DataStorage.checkLocalStorageExist() &&
+        pokemonData.pokemonsDataArr.length > 0
+      )
         localStorage.setItem(
           POKEMONS_LIST_KEY,
           JSON.stringify(pokemonData.pokemonsDataArr)
         );
+      // DataStorage.removeLocalStorage();
     });
   }
